implement create assistant mode
